Update the code to build with a modern Java.
This commit allows the code to compile with any modern java. I tried
with version 11 and 23, which is the latest at the moment, and much
beyond the old obsolete 1.6 that Mauve was using.

Changes include:

In build.xml:
- Set release to 11
- Add debug info
- Remove obsolete (and now broken) rmic dependence
(and a few related changes and formatting changes too)

Java files:
- Fix old-style  new Integer(x)  and so on for the modern  Integer.valueOf(x)
- Annotate deprecated functions and classes with @Deprecated (avoids warnings)

Mauve.java in particular:
- Comment out the code to check for a new version (it makes no sense now and does not work)
- Fix the check for the JVM version (it was making the program not run)

Now it can be compiled with:
  ant compile
  ant jar
and run with:
  java -jar Mauve.jar